routes on pause, fixing models
