+++ billing/models/__init__.py
@@ -1,0 +1,3 @@
+from .truck import Truck
+from .rate import Rate
+from .provider import Provider
