+++ billing/models/provider.py
@@ -1,0 +1,5 @@
+from app import db
+
+class Provider(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(255))
